#118 - refactor pagination on client side
